requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
coverage==5.5
Django==3.2.7
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-filter==2.4.0
django-reversion==4.0.0
djangorestframework==3.12.4
iniconfig==1.1.1
Markdown==3.3.4
mysqlclient==2.0.3
packaging==21.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pydotplus==2.0.2
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
validate-docbr==1.8.2
